Removed frightening error
Previously when the script would attempt to download a Snapchat that
had already been downloaded, it would throw the error “FAILED.” This
seems to be unnecessary and therefore the message has been removed.

A comment was also added to change the USERNAME and PASSWORD fields to
those of your account.